backend/requirements.txt
# ProofFund - Backend
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
httpx==0.27.2
numpy==1.26.4
scikit-learn==1.5.1
openai==1.35.0
google-generativeai>=0.8.0
Pillow>=10.0.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
